relocate the source code before break down the vm_manager.rb file.
