+++ src/bosh_azure_cpi/lib/cloud/azure.rb
@@ -45,8 +45,8 @@ require 'bosh/cpi'
 require 'bosh/cpi/registry_client'
 
 require 'cloud'
-require 'cloud/azure/helpers'
-require 'cloud/azure/logger'
+require 'cloud/azure/utils/helpers'
+require 'cloud/azure/logging/logger'
 require 'cloud/azure/models/config'
 require 'cloud/azure/models/config_factory'
 require 'cloud/azure/models/error_msg'
@@ -57,27 +57,27 @@ require 'cloud/azure/models/instance_id'
 require 'cloud/azure/models/props_factory'
 require 'cloud/azure/models/vm_cloud_props'
 require 'cloud/azure/cloud'
-require 'cloud/azure/azure_client'
+require 'cloud/azure/restapi/azure_client'
 
-require 'cloud/azure/network'
-require 'cloud/azure/manual_network'
-require 'cloud/azure/dynamic_network'
-require 'cloud/azure/vip_network'
-require 'cloud/azure/network_configurator'
-require 'cloud/azure/vm_manager'
-require 'cloud/azure/blob_manager'
-require 'cloud/azure/disk_manager'
-require 'cloud/azure/stemcell_manager'
+require 'cloud/azure/network/network'
+require 'cloud/azure/network/manual_network'
+require 'cloud/azure/network/dynamic_network'
+require 'cloud/azure/network/vip_network'
+require 'cloud/azure/network/network_configurator'
+require 'cloud/azure/vms/vm_manager'
+require 'cloud/azure/storage/blob_manager'
+require 'cloud/azure/disk/disk_manager'
+require 'cloud/azure/stemcell/stemcell_manager'
 require 'cloud/azure/meta_store/meta_store'
 require 'cloud/azure/meta_store/stemcell_meta'
 require 'cloud/azure/meta_store/table_manager'
-require 'cloud/azure/light_stemcell_manager'
+require 'cloud/azure/stemcell/light_stemcell_manager'
 
-require 'cloud/azure/disk_manager2'
-require 'cloud/azure/stemcell_manager2'
-require 'cloud/azure/storage_account_manager'
+require 'cloud/azure/disk/disk_manager2'
+require 'cloud/azure/stemcell/stemcell_manager2'
+require 'cloud/azure/storage/storage_account_manager'
 
-require 'cloud/azure/instance_type_mapper'
+require 'cloud/azure/vms/instance_type_mapper'
 
 require 'cloud/azure/version'
 require 'cloud/azure/telemetry/telemetry_manager'
