+++ src/bosh_azure_cpi/lib/cloud/azure/disk/disk_manager2.rb
