+++ src/bosh_azure_cpi/lib/cloud/azure/logging/logger.rb
