+++ src/bosh_azure_cpi/lib/cloud/azure/network/dynamic_network.rb
