+++ src/bosh_azure_cpi/lib/cloud/azure/network/manual_network.rb
