+++ src/bosh_azure_cpi/lib/cloud/azure/network/network.rb
