+++ src/bosh_azure_cpi/lib/cloud/azure/network/vip_network.rb
