+++ src/bosh_azure_cpi/lib/cloud/azure/restapi/azure_client.rb
