+++ src/bosh_azure_cpi/lib/cloud/azure/stemcell/light_stemcell_manager.rb
