+++ src/bosh_azure_cpi/lib/cloud/azure/stemcell/stemcell_manager.rb
