+++ src/bosh_azure_cpi/lib/cloud/azure/stemcell/stemcell_manager2.rb
