+++ src/bosh_azure_cpi/lib/cloud/azure/storage/blob_manager.rb
