+++ src/bosh_azure_cpi/lib/cloud/azure/storage/storage_account_manager.rb
