+++ src/bosh_azure_cpi/lib/cloud/azure/utils/helpers.rb
