+++ src/bosh_azure_cpi/lib/cloud/azure/vms/instance_type_mapper.rb
