+++ src/bosh_azure_cpi/lib/cloud/azure/vms/vm_manager.rb
